Refactor views to use fewer presenters
