--- app/presenters/find_this_item_presenter.rb
@@ -1,15 +1,0 @@
-# frozen_string_literal: true
-
-class FindThisItemPresenter
-  attr_reader :document, :config
-
-  def initialize(document:)
-    @document = document
-    @config = YAML.safe_load(File.open(Rails.root.join('config', 'metadata', 'find_this_item.yml')))
-  end
-
-  def terms
-    @config = @config.symbolize_keys
-    @document.slice(*@config.keys)
-  end
-end
