--- app/presenters/view_items_in_collection_presenter.rb
@@ -1,15 +1,0 @@
-# frozen_string_literal: true
-
-class ViewItemsInCollectionPresenter
-  attr_reader :document, :config
-
-  def initialize(document:)
-    @document = document
-    @config = YAML.safe_load(File.open(Rails.root.join('config', 'metadata', 'view_items_in_collection.yml')))
-  end
-
-  def terms
-    @config = @config.symbolize_keys
-    @document.slice(*@config.keys)
-  end
-end
