--- app/views/catalog/_about_this_collection.html.erb
@@ -1,1 +1,0 @@
-<%= render 'about_this_collection_block', document: document, presenter: AboutThisCollectionPresenter, presenter_class: 'about-this-collection', title: 'About This Collection' %>
