+++ app/views/catalog/_about_this_collection_block.html.erb
@@ -1,5 +1,5 @@
 <% doc_presenter = show_presenter(document) %>
-<% fields = presenter.new(document: doc_presenter.fields_to_render).terms %>
+<% fields = presenter.new(document: doc_presenter.fields_to_render).terms(:about_this_collection) %>
 <% if fields.length > 0 %>
   <div class="<%= presenter_class %> card mb-2 rounded-0">
     <div class="card-header rounded-0"><%= title %></div>
