--- app/views/catalog/_about_this_item.html.erb
@@ -1,1 +1,0 @@
-<%= render 'metadata_block', document: document, presenter: AboutThisItemPresenter, presenter_class: 'about-this-item', title: 'About This Item', add_class_dt: "col-md-5", add_class_dd: "col-md-7" %>
