--- app/views/catalog/_access_and_copyright.html.erb
@@ -1,1 +1,0 @@
-<%= render 'access_and_copyright_block', document: document, presenter: AccessAndCopyrightPresenter, presenter_class: 'access-and-copyright', title: 'Access and Copyright' %>
