--- app/views/catalog/_additional_details.html.erb
@@ -1,1 +1,0 @@
-<%= render 'metadata_block', document: document, presenter: AdditionalDetailsPresenter, presenter_class: 'additional-details', title: 'Additional Details', add_class_dt: "col-md-5", add_class_dd: "col-md-7" %>
