--- app/views/catalog/_find_this_item.html.erb
@@ -1,1 +1,0 @@
-<%= render 'find_this_item_block', document: document, presenter: FindThisItemPresenter, presenter_class: 'find-this-item', title: 'Find This Item', add_class_dt: "col-md-12", add_class_dd: "col-md-12" %>
