--- app/views/catalog/_is_part_of.html.erb
@@ -1,15 +1,0 @@
-<script>
-
-  // Sends the custom dimension to Google Analytics.
-  gtag('event', "<%= document['member_of_collections_ssim']&.first %>",
-  {
-    'event_category': 'collection_view',
-    'event_label': '<%= document['title_tesim'].first %>'
-  });
-  gtag('event', "<%= document['holding_repository_tesim']&.first %>",
-  {
-    'event_category': 'holding_repository_view',
-    'event_label': '<%= document['title_tesim'].first %>'
-  });
-</script>
-<%= render 'is_part_of_block', document: document, presenter: IsPartOfPresenter, presenter_class: 'is-part-of', title: 'This item is part of:' %>
