+++ app/views/catalog/_metadata_col_2.html.erb
@@ -1,14 +1,12 @@
 <div class="row-flex">
   <% if document["has_model_ssim"]&.first == "Collection" %>
-    <%= render "about_this_collection", document: document %>
+    <%= render 'about_this_collection_block', document: document, presenter: MetadataPresenter, presenter_class: 'about-this-collection', title: 'About This Collection' %>
   <% end %>
   <% if document["has_model_ssim"]&.first == "CurateGenericWork" %>
-    <%= render "this_item_contains", document: document %>
+    <%= render 'this_item_contains_block', document: document, presenter: ThisItemContainsPresenter, presenter_class: 'this-item-contains', title: 'This item contains:' %>
+    <%= render 'metadata_block', document: document, presenter: AboutThisItemPresenter, presenter_class: 'about-this-item', title: 'About This Item', add_class_dt: "col-md-5", add_class_dd: "col-md-7" %>
   <% end %>
-  <% if document["has_model_ssim"]&.first == "CurateGenericWork" %>
-    <%= render "about_this_item", document: document %>
-  <% end %>
-  <%= render "subjects_keywords", document: document %>
-  <%= render "publication_details", document: document %>
-  <%= render "additional_details", document: document %>
+  <%= render 'metadata_block', document: document, presenter: SubjectsKeywordsPresenter, presenter_class: 'subjects-keywords', title: 'Subjects / Keywords', add_class_dt: "col-md-5", add_class_dd: "col-md-7" %>
+  <%= render 'metadata_block', document: document, presenter: PublicationDetailsPresenter, presenter_class: 'publication-details', title: 'Publication Details', add_class_dt: "col-md-5", add_class_dd: "col-md-7" %>
+  <%= render 'metadata_block', document: document, presenter: AdditionalDetailsPresenter, presenter_class: 'additional-details', title: 'Additional Details', add_class_dt: "col-md-5", add_class_dd: "col-md-7" %>
 </div>
