+++ app/views/catalog/_metadata_col_3.html.erb
@@ -1,4 +1,4 @@
 <div class="row-flex">
   <%= render "show_tools", document: document %>
-  <%= render "access_and_copyright", document: document %>
+  <%= render 'access_and_copyright_block', document: document, presenter: MetadataPresenter, presenter_class: 'access-and-copyright', title: 'Access and Copyright' %>
 </div>
