--- app/views/catalog/_publication_details.html.erb
@@ -1,1 +1,0 @@
-<%= render 'metadata_block', document: document, presenter: PublicationDetailsPresenter, presenter_class: 'publication-details', title: 'Publication Details', add_class_dt: "col-md-5", add_class_dd: "col-md-7" %>
