--- app/views/catalog/_related_material.html.erb
@@ -1,1 +1,0 @@
-<%= render 'metadata_block', document: document, presenter: RelatedMaterialPresenter, presenter_class: 'related-material', title: 'Related Material', add_class_dt: "col-md-12", add_class_dd: "col-md-12" %>
