--- app/views/catalog/_subjects_keywords.html.erb
@@ -1,1 +1,0 @@
-<%= render 'metadata_block', document: document, presenter: SubjectsKeywordsPresenter, presenter_class: 'subjects-keywords', title: 'Subjects / Keywords', add_class_dt: "col-md-5", add_class_dd: "col-md-7" %>
