--- app/views/catalog/_this_item_contains.html.erb
@@ -1,1 +1,0 @@
-<%= render 'this_item_contains_block', document: document, presenter: ThisItemContainsPresenter, presenter_class: 'this-item-contains', title: 'This item contains:' %>
