--- app/views/catalog/_view_items_in_collection.html.erb
@@ -1,1 +1,0 @@
-<%= render 'view_items_in_collection_block', document: document, presenter: ViewItemsInCollectionPresenter, presenter_class: 'view-items-in-collection' %>
