+++ app/views/catalog/_view_items_in_collection_block.html.erb
@@ -1,5 +1,5 @@
 <% doc_presenter = show_presenter(document) %>
-<% fields = presenter.new(document: doc_presenter.fields_to_render).terms %>
+<% fields = presenter.new(document: doc_presenter.fields_to_render).terms(:view_items_in_this_collection) %>
   <div class="<%= presenter_class %> card mb-2 rounded-0">
     <div class="card-body">
       <dl>
