+++ config/metadata/metadata.yml
@@ -1,0 +1,29 @@
+about_this_collection:
+  abstract_tesim: 'Description'
+  finding_aid_link_ssm: 'Learn More'
+  creator_tesim: 'Creator'
+  contributors_tesim: 'Contributor'
+  primary_language_tesim: 'Primary Language'
+  notes_tesim: 'Note'
+view_items_in_this_collection:
+  title_tesim:
+is_part_of:
+  member_of_collections_ssim:
+  member_of_collection_ids_ssim:
+  parent_work_for_lux_tesim:
+find_this_item:
+  id: 'Persistent URL'
+  system_of_record_ID_tesim: 'System of Record ID'
+  emory_ark_tesim: 'Emory ARK'
+  other_identifiers_tesim: 'Other Identifiers'
+  institution_tesim: 'Institution'
+  holding_repository_tesim: 'Library'
+  administrative_unit_tesim: 'Administrative Unit'
+  sublocation_tesim: 'Sublocation'
+  local_call_number_tesim: 'Call Number'
+  contact_information_tesim: 'Contact Information'
+access_and_copyright:
+  rights_holders_tesim: 'Rights Holder'
+  human_readable_copyright_date_tesim: 'Copyright Date'
+  human_readable_re_use_license_tesim: 'Re-Use License'
+  access_restriction_notes_tesim: 'Access Restrictions'
